add pagination

+++ comments/views.py
@@ -1,3 +1,4 @@
+from django.contrib import messages
 from django.shortcuts import render, redirect, get_object_or_404
 from django.urls import reverse
 from languages.models import Language
@@ -26,5 +27,9 @@ def delete_comment(request, language_id, comment_id):
     comment = get_object_or_404(Comments, id=comment_id)
     language = Language.objects.get(id=language_id)
     category_id = language.category.id
-    comment.delete()
+    if request.user == comment.user:
+        comment.delete()
+        messages.success(request, 'Отзыв успешно удален.')
+    else:
+        messages.error(request, 'У вас нет прав для удаления этого отзыва.')
     return redirect('languages:language', category_id=category_id, language_id=language_id)
